rename `ray_worker.py`
